+++ packages/puppeteer-extra-plugin-adblocker/ava.config-ts.js
@@ -2,7 +2,7 @@ export default {
   environmentVariables: {
     TS_NODE_COMPILER_OPTIONS: '{"module":"commonjs"}'
   },
-  files: ['src/**.test.ts'],
+  files: ['test/**.test.ts'],
   extensions: ['ts'],
   require: ['ts-node/register']
 }
