+++ packages/puppeteer-extra-plugin-anonymize-ua/src/index.ts
