+++ packages/puppeteer-extra-plugin-anonymize-ua/test/headless.ts
@@ -1,5 +1,3 @@
-'use strict'
-
 import test from 'ava'
 
 const PUPPETEER_ARGS: string[] = ['--no-sandbox', '--disable-setuid-sandbox']
