+++ packages/puppeteer-extra-plugin-anonymize-ua/test/index.test.ts
@@ -1,10 +1,8 @@
-'use strict'
-
 const PLUGIN_NAME = 'anonymize-ua'
 
 import test from 'ava'
 
-import Plugin from '.'
+import Plugin from '../src'
 
 test('is a function', async t => {
   t.is(typeof Plugin, 'function')
