+++ packages/puppeteer-extra-plugin-anonymize-ua/tsconfig.json
