+++ packages/puppeteer-extra-plugin-click-and-wait/src/example.ts
