+++ packages/puppeteer-extra-plugin-click-and-wait/tsconfig.json
