+++ packages/puppeteer-extra-plugin-devtools/test/headless.ts
@@ -1,6 +1,4 @@
-'use strict'
-
-const test = require('ava')
+import test from 'ava'
 
 const PUPPETEER_ARGS = ['--no-sandbox', '--disable-setuid-sandbox']
