+++ packages/puppeteer-extra-plugin/tsconfig.json
@@ -24,5 +24,5 @@
     "./src/**/*.ts",
     "./test/**/*.ts"
   ],
-  "exclude": ["node_modules", "dist", "./test/**/*.spec.ts"]
+  "exclude": ["node_modules", "dist", "./test/**/*.ts"]
 }
